🧹 Change oscPort type from double to int in MainViewModel
The oscPort property was incorrectly typed as a double, which is semantically incorrect for a network port and required an explicit cast when starting the OSC service. This change updates it to an int and removes the redundant cast.

+++ AVP/ViewModels/MainViewModel.cs
@@ -24,7 +24,7 @@ public partial class MainViewModel : ObservableObject
     private string mediaPath = string.Empty;
 
     [ObservableProperty]
-    private double oscPort = 8000;
+    private int oscPort = 8000;
 
     [ObservableProperty]
     private bool isOscRunning;
@@ -76,7 +76,7 @@ public partial class MainViewModel : ObservableObject
     [RelayCommand(CanExecute = nameof(CanStartOsc))]
     private void StartOsc()
     {
-        if (_oscService.Start((int)OscPort))
+        if (_oscService.Start(OscPort))
         {
             IsOscRunning = true;
             StartOscCommand.NotifyCanExecuteChanged();
